refactored internals for reusability
this will be used for upcoming virtual bundles
