+++ lib/bundle/config.js
@@ -54,8 +54,8 @@ function generateConfig({ extensions = [], // eslint-disable-next-line indent
 		esnext, jsx, typescript, // eslint-disable-next-line indent
 		sourcemap, compact }, { browsers }) {
 	let cfg = { sourcemap };
-
 	let plugins = [];
+
 	if(esnext || jsx) {
 		let transpiler = Object.assign({}, esnext, jsx);
 		if(esnext) {
