+++ lib/bundle/index.js
@@ -1,24 +1,22 @@
 "use strict";
 
-let generateBundle = require("./bundler");
-let generateConfig = require("./config");
-let { generateError } = require("../util");
+let BasicBundle = require("./basic");
 
-let DEFAULTS = {
-	format: "iife"
-};
-
-module.exports = class Bundle {
+module.exports = class Bundle extends BasicBundle {
 	constructor(entryPoint, target, config, { browsers }) {
-		this.entryPoint = entryPoint;
-		this.target = target;
-
-		config = Object.assign({}, DEFAULTS, config);
+		// extract bundle-specific fingerprinting, if any
+		config = Object.assign({}, config);
 		if(config.fingerprint !== undefined) {
-			this.fingerprint = config.fingerprint;
+			var fingerprint = config.fingerprint; // eslint-disable-line no-var
 			delete config.fingerprint;
 		}
-		this._config = generateConfig(config, { browsers });
+		super(config, { browsers });
+
+		this.entryPoint = entryPoint;
+		this.target = target;
+		if(fingerprint !== undefined) {
+			this.fingerprint = fingerprint;
+		}
 	}
 
 	// compiles the bundle - if a list of file paths is provided, compilation
@@ -32,15 +30,6 @@ module.exports = class Bundle {
 			return false;
 		}
 
-		return generateBundle(this.entryPoint, this._config, this._cache).
-			then(({ code, modules, cache }) => {
-				this._modules = modules;
-				this._cache = cache;
-				return { code };
-			}, err => ({
-				// also report error from within bundle, to avoid it being overlooked
-				code: generateError(err),
-				error: err
-			}));
+		return super.compile(this.entryPoint);
 	}
 };
